jwt access/refresh, fix bugs

+++ frontend/src/app/components/comment-list/comment-card.ts
@@ -2,6 +2,7 @@ import { Component, ElementRef, EventEmitter, inject, Input, Output, ViewChild }
 import { RouterLink } from '@angular/router';
 import { CommentResponseWrapper } from '../../types/comments';
 import { ProfileService } from '../../services/profile.service';
+import { JwtService } from '../../services/jwt.service';
 
 @Component({
   selector: 'app-comment-card',
@@ -10,6 +11,7 @@ import { ProfileService } from '../../services/profile.service';
 })
 export class CommentCard {
   private profileService = inject(ProfileService)
+  private jwtService = inject(JwtService)
 
   @Input({ required: true }) comment!: CommentResponseWrapper
   @Input({ required: true }) concertSlug!: string
@@ -18,12 +20,20 @@ export class CommentCard {
   @ViewChild('settings') settingsDialog!: ElementRef<HTMLDetailsElement>
 
   toggleCommentLike() {
+    if(this.jwtService.checkIfLogged()) {
+      return
+    }
+
     this.profileService.setLikeComment(this.comment.id, !this.comment.liked).subscribe(() => {
       this.comment.liked = !this.comment.liked
     })
   }
 
   settingsDeleteComment() {
+    if(this.jwtService.checkIfLogged()) {
+      return
+    }
+
     this.profileService.deleteComment(this.concertSlug, this.comment.id).subscribe(() => {
       this.settingsDialog.nativeElement.open = false
       this.onDelete.emit(this.comment)
